refactor: re-export PagingStrategy type from paging strategies index.ts
