+++ src/hooks/index.ts
@@ -6,9 +6,11 @@ export * from './carousel/useCarouselTile';
 export * from './layout/useScrolledToEdge';
 export * from './layout/useTargetZone';
 
+export * from './paging/strategies';
 export * from './paging/usePaging';
 export * from './paging/usePageWithArrowKeys';
-export * from './paging/strategies';
+export * from './paging/usePages';
+export * from './paging/usePageIndicator';
 
 export * from './scroll/useKeepTileInView';
 export * from './scroll/useScrollFocusedIntoView';
@@ -18,5 +20,3 @@ export * from './scroll/useScrollSnapLoadingFix';
 export * from './select/useSelectOnFocus';
 export * from './select/useSelectWithArrowKeys';
 
-export * from './paging/usePages';
-export * from './paging/usePageIndicator';
