+++ src/hooks/paging/strategies/index.ts
@@ -1,1 +1,2 @@
+export type { PagingStrategy } from './common';
 export { pageByVisibility } from './pageByVisibility';
